posts/requirements.txt
asyncpg
databases[postgresql]
fastapi
SQLAlchemy
uvicorn
pydantic
psycopg2-binary
regex
python-jose
bcrypt==4.0.1
passlib
python-jose
alembic
pydantic-settings
pydantic[email]
grpcio==1.71.0
grpcio-tools